requirements.txt
# Core LLM and Agent Framework
langchain==0.3.13
langchain-anthropic==0.3.3
langchain-core==0.3.28
langgraph==0.2.60

# Utilities
python-dotenv==1.0.0
pydantic==2.10.5

# Optional: For future RAG integration
# langchain-community==0.3.13
# chromadb==0.5.23
# sentence-transformers==3.3.1

# Optional: For web UI
# streamlit==1.41.1
# gradio==5.10.0

# Development
pytest==8.3.4
black==24.10.0
